requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
pypdf2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
aiofiles==23.2.1
lxml==4.9.3
google-generativeai==0.3.2
mistralai==1.0.0
scikit-learn
numpy
aiohttp
